e2e: remove unused framework provisioning code
