+++ e2e/framework/case.go
@@ -5,7 +5,6 @@ import (
 
 	capi "github.com/hashicorp/consul/api"
 	"github.com/hashicorp/nomad/api"
-	"github.com/hashicorp/nomad/e2e/framework/provisioning"
 )
 
 // TestSuite defines a set of test cases and under what conditions to run them
@@ -59,7 +58,7 @@ func (c Constraints) matches(env Environment) error {
 
 // TC is the base test case which should be embedded in TestCase implementations.
 type TC struct {
-	cluster *provisioning.ClusterInfo
+	cluster *ClusterInfo
 }
 
 // Nomad returns a configured nomad api client
@@ -78,6 +77,6 @@ func (tc *TC) Name() string {
 	return tc.cluster.Name
 }
 
-func (tc *TC) setClusterInfo(info *provisioning.ClusterInfo) {
+func (tc *TC) setClusterInfo(info *ClusterInfo) {
 	tc.cluster = info
 }
