+++ e2e/framework/framework.go
@@ -7,8 +7,6 @@ import (
 	"reflect"
 	"strings"
 	"testing"
-
-	"github.com/hashicorp/nomad/e2e/framework/provisioning"
 )
 
 const frameworkHelp = `
@@ -17,28 +15,10 @@ Usage: go test -v ./e2e [options]
 These flags are coarse overrides for the test environment.
 
   -forceRun    skip all environment checks when filtering test suites
-  -local       force default no-op provisioning
-  -skipTests   skips all tests and only provisions
+  -local       denotes execution is against a local environment
   -slow        include execution of slow test suites
   -suite       run specified test suite
   -showHelp    shows this help text
-
-Provisioning flags tell the test runner to pre-provision the cluster before
-running all tests. These flags can be passed to 'go test'. If no
-'-provision.*' flag is set, the test runner assumes the cluster has already
-been configured and uses the test environment's env vars to connect to the
-cluster.
-
-  -provision.terraform=string   pass file generated by terraform output
-  -provision.vagrant=string     provision to a single-node vagrant box
-
-Nomad version flags tell the provisioner to deploy a specific version of
-Nomad. These flags are all ignored if no '-provision.*' flag is set.
-Otherwise at most one should be set.
-
-  -nomad.local_file=string  provision this specific local binary of Nomad
-  -nomad.sha=string         provision this specific sha from S3
-  -nomad.version=string     provision this version from releases.hashicorp.com
 
 TestSuites can request Constraints on the Framework.Environment so that tests
 are only run in the appropriate conditions. These environment flags provide
@@ -54,35 +34,17 @@ the information for those constraints.
 
 var fHelp = flag.Bool("showHelp", false, "print the help screen")
 var fLocal = flag.Bool("local", false,
-	"denotes execution is against a local environment, forcing default no-op provisioning")
+	"denotes execution is against a local environment")
 var fSlow = flag.Bool("slow", false, "toggles execution of slow test suites")
 var fForceRun = flag.Bool("forceRun", false,
 	"if set, skips all environment checks when filtering test suites")
-var fSkipTests = flag.Bool("skipTests", false, "skip all tests and only provision")
 var fSuite = flag.String("suite", "", "run specified test suite")
-
-// Provisioning flags
-var fProvisionVagrant = flag.String("provision.vagrant", "",
-	"run pre-provision to a single-node vagrant host")
-var fProvisionTerraform = flag.String("provision.terraform", "",
-	"run pre-provision from file generated by 'terraform output provisioning'")
-
-// Nomad version flags
-// TODO: these override each other. local_file > sha > version
-// but we should assert at most 1 is set.
-var fProvisionNomadLocalBinary = flag.String("nomad.local_file", "",
-	"provision this specific local binary of Nomad (ignored for no-op provisioning).")
-var fProvisionNomadSha = flag.String("nomad.sha", "",
-	"provision this specific sha of Nomad (ignored for no-op provisioning)")
-var fProvisionNomadVersion = flag.String("nomad.version", "",
-	"provision this specific release of Nomad (ignored for no-op provisioning)")
 
 // Environment flags
 // TODO:
-// if we have a provisioner, each target has its own environment. it'd
-// be nice if we could match that environment against the tests so that
-// we always avoid running tests that don't apply against the
-// environment, and then have these flags override that behavior.
+// It would be nice if we could match the target environment against
+// the tests automatically so that we always avoid running tests that
+// don't apply, and then have these flags override that behavior.
 var fEnv = flag.String("env", "", "name of the environment executing against")
 var fProvider = flag.String("env.provider", "",
 	"cloud provider for which environment is executing against")
@@ -95,13 +57,12 @@ var fTags = flag.String("env.tags", "",
 
 type Framework struct {
 	suites      []*TestSuite
-	provisioner provisioning.Provisioner
+	provisioner Provisioner
 	env         Environment
 
 	isLocalRun bool
 	slow       bool
 	force      bool
-	skipAll    bool
 	suite      string
 }
 
@@ -132,20 +93,12 @@ func New() *Framework {
 		env.Tags[tag] = struct{}{}
 	}
 	return &Framework{
-		provisioner: provisioning.NewProvisioner(provisioning.ProvisionerConfig{
-			IsLocal:          *fLocal,
-			VagrantBox:       *fProvisionVagrant,
-			TerraformConfig:  *fProvisionTerraform,
-			NomadLocalBinary: *fProvisionNomadLocalBinary,
-			NomadSha:         *fProvisionNomadSha,
-			NomadVersion:     *fProvisionNomadVersion,
-		}),
-		env:        env,
-		isLocalRun: *fLocal,
-		slow:       *fSlow,
-		force:      *fForceRun,
-		skipAll:    *fSkipTests,
-		suite:      *fSuite,
+		provisioner: DefaultProvisioner,
+		env:         env,
+		isLocalRun:  *fLocal,
+		slow:        *fSlow,
+		force:       *fForceRun,
+		suite:       *fSuite,
 	}
 }
 
@@ -164,15 +117,11 @@ func AddSuites(s ...*TestSuite) {
 
 // Run starts the test framework, running each TestSuite
 func (f *Framework) Run(t *testing.T) {
-	info, err := f.provisioner.SetupTestRun(t, provisioning.SetupOptions{})
+	info, err := f.provisioner.SetupTestRun(t, SetupOptions{})
 	if err != nil {
 		t.Fatalf("could not provision cluster: %v", err)
 	}
 	defer f.provisioner.TearDownTestRun(t, info.ID)
-
-	if f.skipAll {
-		t.Skip("Skipping all tests, -skipTests set")
-	}
 
 	for _, s := range f.suites {
 		t.Run(s.Component, func(t *testing.T) {
@@ -224,7 +173,7 @@ func (f *Framework) runSuite(t *testing.T, s *TestSuite) (skip bool, err error) 
 		return true, fmt.Errorf("only running suite %q", f.suite)
 	}
 
-	info, err := f.provisioner.SetupTestSuite(t, provisioning.SetupOptions{
+	info, err := f.provisioner.SetupTestSuite(t, SetupOptions{
 		Name:         s.Component,
 		ExpectConsul: s.Consul,
 		ExpectVault:  s.Vault,
@@ -248,7 +197,7 @@ func (f *Framework) runCase(t *testing.T, s *TestSuite, c TestCase) {
 
 	// The ClusterInfo handle should be used by each TestCase to isolate
 	// job/task state created during the test.
-	info, err := f.provisioner.SetupTestCase(t, provisioning.SetupOptions{
+	info, err := f.provisioner.SetupTestCase(t, SetupOptions{
 		Name:         name,
 		ExpectConsul: s.Consul,
 		ExpectVault:  s.Vault,
