+++ e2e/framework/interfaces.go
@@ -1,7 +1,5 @@
 package framework
 
-import "github.com/hashicorp/nomad/e2e/framework/provisioning"
-
 // TestCase is the interface which an E2E test case implements.
 // It is not meant to be implemented directly, instead the struct should embed
 // the 'framework.TC' struct
@@ -12,7 +10,7 @@ type TestCase interface {
 }
 
 type internalTestCase interface {
-	setClusterInfo(*provisioning.ClusterInfo)
+	setClusterInfo(*ClusterInfo)
 }
 
 // BeforeAllTests is used to define a method to be called before the execution
@@ -32,7 +30,7 @@ type BeforeEachTest interface {
 	BeforeEach(*F)
 }
 
-// AfterEachTest is used to degine a method to be called after each test.
+// AfterEachTest is used to define a method to be called after each test.
 type AfterEachTest interface {
 	AfterEach(*F)
 }
