--- e2e/framework/provisioning/example.json
@@ -1,52 +1,0 @@
-{
-  "clients": [
-    {
-      "deployment": {
-        "bundles": [
-          {
-            "destination": "/ops/shared",
-            "source": "/home/me/nomad_src/e2e/terraform/shared"
-          }
-        ],
-        "nomad_sha": "126b86557353054cc40db474a688a27281aa85e6",
-        "platform": "linux_amd64",
-        "remote_binary_path": "/usr/local/bin/nomad",
-        "steps": [
-          "sudo chmod +x /ops/shared/provision-client.sh",
-          "sudo /ops/shared/config/provision-client.sh aws '126b86557353054cc40db474a688a27281aa85e6' 'indexed/client-0.hcl'"
-        ]
-      },
-      "runner": {
-        "host": "192.168.1.101",
-        "key": "/home/me/nomad_src/e2e/terraform/keys/random-animal.pem",
-        "port": 22,
-        "user": "ubuntu"
-      }
-    }
-  ],
-  "servers": [
-    {
-      "deployment": {
-        "bundles": [
-          {
-            "destination": "/ops/shared",
-            "source": "/home/me/nomad_src/e2e/terraform/shared"
-          }
-        ],
-        "nomad_sha": "126b86557353054cc40db474a688a27281aa85e6",
-        "platform": "linux_amd64",
-        "remote_binary_path": "/usr/local/bin/nomad",
-        "steps": [
-          "sudo chmod +x /ops/shared/config/provision-server.sh",
-          "sudo /ops/shared/config/provision-server.sh aws 1 '126b86557353054cc40db474a688a27281aa85e6' 'indexed/server-0.hcl'"
-        ]
-      },
-      "runner": {
-        "host": "192.168.1.102",
-        "key": "/home/me/nomad_src/e2e/terraform/keys/random-animal.pem",
-        "port": 22,
-        "user": "ubuntu"
-      }
-    }
-  ]
-}
